test: add negative tests
