+++ actix-web-codegen/tests/trybuild/scope-missing-args.rs
@@ -1,0 +1,6 @@
+use actix_web_codegen::scope;
+
+#[scope]
+mod api {}
+
+fn main() {}
