+++ actix-web-codegen/tests/trybuild/scope-on-handler.rs
@@ -1,0 +1,8 @@
+use actix_web_codegen::scope;
+
+#[scope("/api")]
+async fn index() -> &'static str {
+    "Hello World!"
+}
+
+fn main() {}
